Rename: "파일 구조 변경"
todo와 관련된 파일을 todo폴더에 정리해줌.
